+++ src/actions/index.js
@@ -11,6 +11,7 @@ export {
   deleteContent,
   editContent,
   getContent,
+  orderContent,
 } from './content/content';
 export getNavigation from './navigation/navigation';
 export getSchema from './schema/schema';